feat(compact): Add Pushgateway metrics support for one-shot runs
This change introduces the ability for the `thanos compact` command to
push its metrics to a Prometheus Pushgateway when running in one-shot
mode (without the `--wait` flag).

This is particularly useful for monitoring the compactor when it is run
as a periodic batch job, such as a cron job.

The following changes have been made:
- Refactored progress metric calculation to run in one-shot mode.
- Added `--push-gateway.url` and `--push-gateway.job` flags to configure
the Pushgateway endpoint and job name.
- Implemented the logic to push all registered metrics upon successful
completion of a one-shot compaction run.

Signed-off-by: Enno Richter <enno@nerdworks.de>